OneDrive/Documents/Desktop/chatbot-AI/requirements.txt
Flask==3.0.2

Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.10
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1